+++ threedigrid_builder/lib/nodes.pyx
@@ -1,8 +1,8 @@
-cimport nodes
-cimport quadtree
+from . cimport nodes
+from . cimport quadtree
 import numpy as np
 from numpy cimport npy_intp, ndarray, NPY_DOUBLE, NPY_INT
-from utils cimport create_array
+from .utils cimport create_array
 
 
 cdef class Nodes: