+++ CHANGES.rst
@@ -1,5 +1,7 @@
-Changelog of py3di
-==================
+Changelog of threedigrid-builder
+================================
 
 0.1 (unreleased)
 ----------------
+
+- Nothing added yet.